+++ CHANGELOG.md
@@ -15,6 +15,7 @@ phpSysInfo 3.2.9
  - [NEW] Thinkpad - Lenovo Thinkpad sensors monitoring for Linux
  - [NEW] StableBit plugin - StableBit Scanner info
  - [NEW] HyperV plugin - Hyper-V info
+ - [NEW] Raid plugin - Software, Fake and Zpool RAID status
 
 phpSysInfo 3.2.8
 ----------------